refactor: move fetchGraphRelated to db layer, fix sanitized query usage
- Move fetchGraphRelated() and GraphRelatedThought from src/tools/search.ts
  to src/db/edges.ts where AGENTS.md says all SQLite operations belong
- Pass sanitized query instead of raw searchText to searchThoughts in
  capture.ts findSuggestedConnections()
