+++ src/db/edges.ts
@@ -247,6 +247,102 @@ export function getConnections(
  * BFS traversal from a starting thought up to max_depth (capped at 5).
  * Returns all discovered nodes with their depth and edges.
  */
+export interface GraphRelatedThought {
+  id: string;
+  summary: string | null;
+  type: string;
+  depth: number;
+  via_edge_type: string;
+  direction: "outgoing" | "incoming";
+}
+
+/**
+ * After retrieving a set of result IDs, traverse their graph edges up to
+ * graphDepth hops and return related thoughts not already in the result set.
+ */
+export function fetchGraphRelated(
+  db: ThoughtDatabase,
+  resultIds: string[],
+  graphDepth: number,
+): GraphRelatedThought[] {
+  if (graphDepth <= 0 || resultIds.length === 0) return [];
+
+  const effectiveDepth = Math.min(Math.max(graphDepth, 1), 5);
+  const seen = new Set<string>(resultIds);
+  const related: GraphRelatedThought[] = [];
+
+  // BFS from all result nodes simultaneously
+  type QueueItem = { id: string; depth: number };
+  const queue: QueueItem[] = resultIds.map((id) => ({ id, depth: 0 }));
+
+  while (queue.length > 0) {
+    const current = queue.shift()!;
+    if (current.depth >= effectiveDepth) continue;
+
+    // Outgoing edges
+    const outgoing = db.db
+      .prepare(
+        `SELECT e.edge_type, e.target_id, t.summary, t.type
+         FROM edges e
+         JOIN thoughts t ON t.id = e.target_id
+         WHERE e.source_id = ?`,
+      )
+      .all(current.id) as Array<{
+      edge_type: string;
+      target_id: string;
+      summary: string | null;
+      type: string;
+    }>;
+
+    for (const row of outgoing) {
+      if (!seen.has(row.target_id)) {
+        seen.add(row.target_id);
+        related.push({
+          id: row.target_id,
+          summary: row.summary,
+          type: row.type,
+          depth: current.depth + 1,
+          via_edge_type: row.edge_type,
+          direction: "outgoing",
+        });
+        queue.push({ id: row.target_id, depth: current.depth + 1 });
+      }
+    }
+
+    // Incoming edges
+    const incoming = db.db
+      .prepare(
+        `SELECT e.edge_type, e.source_id, t.summary, t.type
+         FROM edges e
+         JOIN thoughts t ON t.id = e.source_id
+         WHERE e.target_id = ?`,
+      )
+      .all(current.id) as Array<{
+      edge_type: string;
+      source_id: string;
+      summary: string | null;
+      type: string;
+    }>;
+
+    for (const row of incoming) {
+      if (!seen.has(row.source_id)) {
+        seen.add(row.source_id);
+        related.push({
+          id: row.source_id,
+          summary: row.summary,
+          type: row.type,
+          depth: current.depth + 1,
+          via_edge_type: row.edge_type,
+          direction: "incoming",
+        });
+        queue.push({ id: row.source_id, depth: current.depth + 1 });
+      }
+    }
+  }
+
+  return related;
+}
+
 export function traverseGraph(
   db: ThoughtDatabase,
   thought_id: string,
